+++ src/main/kotlin/dev/aaronhowser/mods/irregular_implements/block/block_entity/ItemCollectorBlockEntity.kt
@@ -53,7 +53,6 @@ open class ItemCollectorBlockEntity(
 
 			itemEntity.item = newStack
 		}
-
 	}
 
 	open fun getFilter(): ItemFilterDataComponent? {